BufferingStream -> BufferingWriter
