+++ src/Http/Http/src/PublicAPI.Unshipped.txt
@@ -1,4 +1,4 @@
 #nullable enable
 *REMOVED*Microsoft.AspNetCore.Http.StreamResponseBodyFeature.StreamResponseBodyFeature(System.IO.Stream! stream, Microsoft.AspNetCore.Http.Features.IHttpResponseBodyFeature! priorFeature) -> void
-Microsoft.AspNetCore.Http.StreamResponseBodyFeature.FlushWriterAsync() -> System.Threading.Tasks.Task!
+Microsoft.AspNetCore.Http.StreamResponseBodyFeature.CompleteWriterAsync() -> System.Threading.Tasks.Task!
 Microsoft.AspNetCore.Http.StreamResponseBodyFeature.StreamResponseBodyFeature(System.IO.Stream! stream, Microsoft.AspNetCore.Http.Features.IHttpResponseBodyFeature? priorFeature) -> void
