+++ src/Middleware/HttpLogging/src/ResponseBufferingStream.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Buffers;
+using System.IO.Pipelines;
 using System.Text;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
@@ -13,15 +14,21 @@ namespace Microsoft.AspNetCore.HttpLogging;
 /// <summary>
 /// Stream that buffers reads
 /// </summary>
-internal sealed class ResponseBufferingStream : BufferingStream
+internal sealed class ResponseBufferingStream : Stream, IHttpResponseBodyFeature
 {
     private readonly IHttpResponseBodyFeature _innerBodyFeature;
+    private readonly Stream _innerStream;
     private readonly int _limit;
+    private PipeWriter? _pipeAdapter;
+    private readonly BufferingWriter _bufferingWriter;
 
     private readonly HttpContext _context;
     private readonly List<MediaTypeState> _encodings;
     private readonly HttpLoggingOptions _options;
     private Encoding? _encoding;
+    private readonly ILogger _logger;
+
+    private static readonly StreamPipeWriterOptions _pipeWriterOptions = new StreamPipeWriterOptions(leaveOpen: true);
 
     internal ResponseBufferingStream(IHttpResponseBodyFeature innerBodyFeature,
         int limit,
@@ -29,7 +36,6 @@ internal sealed class ResponseBufferingStream : BufferingStream
         HttpContext context,
         List<MediaTypeState> encodings,
         HttpLoggingOptions options)
-        : base(innerBodyFeature.Stream, logger)
     {
         _innerBodyFeature = innerBodyFeature;
         _innerStream = innerBodyFeature.Stream;
@@ -37,9 +43,97 @@ internal sealed class ResponseBufferingStream : BufferingStream
         _context = context;
         _encodings = encodings;
         _options = options;
+        _logger = logger;
+        _bufferingWriter = new BufferingWriter();
+    }
+
+    public override bool CanSeek => _innerStream.CanSeek;
+
+    public override bool CanRead => _innerStream.CanRead;
+
+    public override bool CanWrite => _innerStream.CanWrite;
+
+    public override long Length => _innerStream.Length;
+
+    public override long Position
+    {
+        get => _innerStream.Position;
+        set => _innerStream.Position = value;
+    }
+
+    public override int WriteTimeout
+    {
+        get => _innerStream.WriteTimeout;
+        set => _innerStream.WriteTimeout = value;
+    }
+
+    public string GetString(Encoding? encoding)
+    {
+        try
+        {
+            if (_bufferingWriter.Head == null || _bufferingWriter.Tail == null)
+            {
+                // nothing written
+                return "";
+            }
+
+            if (encoding == null)
+            {
+                _logger.UnrecognizedMediaType();
+                return "";
+            }
+
+            // Only place where we are actually using the buffered data.
+            // update tail here.
+            _bufferingWriter.Tail.End = _bufferingWriter.TailBytesBuffered;
+
+            var ros = new ReadOnlySequence<byte>(_bufferingWriter.Head, 0, _bufferingWriter.Tail, _bufferingWriter.TailBytesBuffered);
+
+            var bufferWriter = new ArrayBufferWriter<char>();
+
+            var decoder = encoding.GetDecoder();
+            // First calls convert on the entire ReadOnlySequence, with flush: false.
+            // flush: false is required as we don't want to write invalid characters that
+            // are spliced due to truncation. If we set flush: true, if effectively means
+            // we expect EOF in this array, meaning it will try to write any bytes at the end of it.
+            EncodingExtensions.Convert(decoder, ros, bufferWriter, flush: false, out var charUsed, out var completed);
+
+            // Afterwards, we need to call convert in a loop until complete is true.
+            // The first call to convert many return true, but if it doesn't, we call
+            // Convert with a empty ReadOnlySequence and flush: true until we get completed: true.
+
+            // This should never infinite due to the contract for decoders.
+            // But for safety, call this only 10 times, throwing a decode failure if it fails.
+            for (var i = 0; i < 10; i++)
+            {
+                if (completed)
+                {
+                    return new string(bufferWriter.WrittenSpan);
+                }
+                else
+                {
+                    EncodingExtensions.Convert(decoder, ReadOnlySequence<byte>.Empty, bufferWriter, flush: true, out charUsed, out completed);
+                }
+            }
+
+            throw new DecoderFallbackException("Failed to decode after 10 calls to Decoder.Convert");
+        }
+        catch (DecoderFallbackException ex)
+        {
+            _logger.DecodeFailure(ex);
+            return "<Decoder failure>";
+        }
+        finally
+        {
+            Reset();
+        }
     }
 
     public bool FirstWrite { get; private set; }
+
+    public Stream Stream => this;
+
+    public PipeWriter Writer => _pipeAdapter ??= PipeWriter.Create(Stream, _pipeWriterOptions);
 
     public Encoding? Encoding { get => _encoding; }
 
@@ -60,22 +154,22 @@ internal sealed class ResponseBufferingStream : BufferingStream
 
     public override void Write(ReadOnlySpan<byte> span)
     {
-        var remaining = _limit - _bytesBuffered;
+        var remaining = _limit - _bufferingWriter.BytesBuffered;
         var innerCount = Math.Min(remaining, span.Length);
 
         OnFirstWrite();
 
         if (innerCount > 0)
         {
-            if (span.Slice(0, innerCount).TryCopyTo(_tailMemory.Span))
+            if (span.Slice(0, innerCount).TryCopyTo(_bufferingWriter.TailMemory.Span))
             {
-                _tailBytesBuffered += innerCount;
-                _bytesBuffered += innerCount;
-                _tailMemory = _tailMemory.Slice(innerCount);
+                _bufferingWriter.TailBytesBuffered += innerCount;
+                _bufferingWriter.BytesBuffered += innerCount;
+                _bufferingWriter.TailMemory = _bufferingWriter.TailMemory.Slice(innerCount);
             }
             else
             {
-                BuffersExtensions.Write(this, span.Slice(0, innerCount));
+                BuffersExtensions.Write(_bufferingWriter, span.Slice(0, innerCount));
             }
         }
 
@@ -89,23 +183,23 @@ internal sealed class ResponseBufferingStream : BufferingStream
 
     public override async ValueTask WriteAsync(ReadOnlyMemory<byte> buffer, CancellationToken cancellationToken = default)
     {
-        var remaining = _limit - _bytesBuffered;
+        var remaining = _limit - _bufferingWriter.BytesBuffered;
         var innerCount = Math.Min(remaining, buffer.Length);
 
         OnFirstWrite();
 
         if (innerCount > 0)
         {
-            if (_tailMemory.Length - innerCount > 0)
+            if (_bufferingWriter.TailMemory.Length - innerCount > 0)
             {
-                buffer.Slice(0, innerCount).CopyTo(_tailMemory);
-                _tailBytesBuffered += innerCount;
-                _bytesBuffered += innerCount;
-                _tailMemory = _tailMemory.Slice(innerCount);
+                buffer.Slice(0, innerCount).CopyTo(_bufferingWriter.TailMemory);
+                _bufferingWriter.TailBytesBuffered += innerCount;
+                _bufferingWriter.BytesBuffered += innerCount;
+                _bufferingWriter.TailMemory = _bufferingWriter.TailMemory.Slice(innerCount);
             }
             else
             {
-                BuffersExtensions.Write(this, buffer.Span);
+                BuffersExtensions.Write(_bufferingWriter, buffer.Span);
             }
         }
 
@@ -124,15 +218,118 @@ internal sealed class ResponseBufferingStream : BufferingStream
         }
     }
 
+    public void DisableBuffering()
+    {
+        _innerBodyFeature.DisableBuffering();
+    }
+
+    public Task SendFileAsync(string path, long offset, long? count, CancellationToken cancellation)
+    {
+        OnFirstWrite();
+        return _innerBodyFeature.SendFileAsync(path, offset, count, cancellation);
+    }
+
+    public Task StartAsync(CancellationToken token = default)
+    {
+        OnFirstWrite();
+        return _innerBodyFeature.StartAsync(token);
+    }
+
+    public async Task CompleteAsync()
+    {
+        await _innerBodyFeature.CompleteAsync();
+    }
+
+    protected override void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            Reset();
+        }
+    }
+
+    public void Reset()
+    {
+        var segment = _bufferingWriter.Head;
+        while (segment != null)
+        {
+            var returnSegment = segment;
+            segment = segment.NextSegment;
+
+            // We haven't reached the tail of the linked list yet, so we can always return the returnSegment.
+            returnSegment.ResetMemory();
+        }
+
+        _bufferingWriter.Head = _bufferingWriter.Tail = null;
+
+        _bufferingWriter.BytesBuffered = 0;
+        _bufferingWriter.TailBytesBuffered = 0;
+    }
+
     public override void Flush()
     {
         OnFirstWrite();
-        base.Flush();
+        _innerStream.Flush();
     }
 
     public override Task FlushAsync(CancellationToken cancellationToken)
     {
         OnFirstWrite();
-        return base.FlushAsync(cancellationToken);
+        return _innerStream.FlushAsync(cancellationToken);
+    }
+
+    public override int Read(byte[] buffer, int offset, int count)
+    {
+        return _innerStream.Read(buffer, offset, count);
+    }
+
+    public override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
+    {
+        return _innerStream.ReadAsync(buffer, offset, count, cancellationToken);
+    }
+
+    public override long Seek(long offset, SeekOrigin origin)
+    {
+        return _innerStream.Seek(offset, origin);
+    }
+
+    public override void SetLength(long value)
+    {
+        _innerStream.SetLength(value);
+    }
+
+    public override IAsyncResult BeginRead(byte[] buffer, int offset, int count, AsyncCallback? callback, object? state)
+    {
+        return _innerStream.BeginRead(buffer, offset, count, callback, state);
+    }
+
+    public override int EndRead(IAsyncResult asyncResult)
+    {
+        return _innerStream.EndRead(asyncResult);
+    }
+
+    public override void CopyTo(Stream destination, int bufferSize)
+    {
+        _innerStream.CopyTo(destination, bufferSize);
+    }
+
+    public override Task CopyToAsync(Stream destination, int bufferSize, CancellationToken cancellationToken)
+    {
+        return _innerStream.CopyToAsync(destination, bufferSize, cancellationToken);
+    }
+
+    public override ValueTask<int> ReadAsync(Memory<byte> buffer, CancellationToken cancellationToken = default)
+    {
+        return _innerStream.ReadAsync(buffer, cancellationToken);
+    }
+
+    public override ValueTask DisposeAsync()
+    {
+        return _innerStream.DisposeAsync();
+    }
+
+    public override int Read(Span<byte> buffer)
+    {
+        return _innerStream.Read(buffer);
     }
 }
